refactor(utils): extract parseTxHash to utils/parseCliOutput.ts (no behavior change)
The same regex pair for extracting `transaction hash: 0x…` (with
64-char-hex fallback) lived in two places after M2.2:
core/Publisher.ts and harness/codeObject.ts. M2.3 (#118) needs a
third caller in harness/script.ts. Lift once, import three times.

  - New utils/parseCliOutput.ts owns `parseTxHash(stdout)`.
  - Publisher.ts: 13-line inline regex block → 1-line `parseTxHash()`
    call; new import line.
  - harness/codeObject.ts: deletes the local `parseTxHash` function;
    new import line.

Existing tests cover both consumers transitively. Verified
behavior-neutral: pnpm --filter movehat test shows 210/210 unchanged.
tsc --noEmit clean.

Tracks #69. Step 1 of 2 for M2.3 (#118).
